Add files via upload
Upgraded version of my code. I got the price to print out as a double. The arithmetic with the price of the total pizza including topics is now correct. There is now a menu in my code. There is also some pauses in my code so the user is not overwhelmed by too many prompts. Feel pretty good about this one.